refactor: centralize CONTACT_EMAIL in lib/constants.ts
