+++ constants/index.ts
@@ -3,5 +3,3 @@ export const PAGINATION = {
   // Default number of items to display per page and for initial data loading
   DEFAULT_PAGE_SIZE: 25,
 };
-
-export const CONTACT_EMAIL = "support@mietevo.de";
